refactor : front end ui  huge monolithic  to seperate chunks
