+++ src/components/ConfidenceRing.jsx
@@ -1,0 +1,42 @@
+import React from "react";
+
+export default function ConfidenceRing({ value }) {
+  const r = 28;
+  const circ = 2 * Math.PI * r;
+  const dash = (value / 100) * circ;
+
+  return (
+    <div style={{ position: "relative", width: 72, height: 72, flexShrink: 0 }}>
+      <svg width={72} height={72} style={{ transform: "rotate(-90deg)" }}>
+        <circle cx={36} cy={36} r={r} fill="none" stroke="#1e2d3d" strokeWidth={6} />
+        <circle
+          cx={36}
+          cy={36}
+          r={r}
+          fill="none"
+          stroke="#00e5a0"
+          strokeWidth={6}
+          strokeDasharray={`${dash} ${circ - dash}`}
+          strokeLinecap="round"
+          style={{ transition: "stroke-dasharray 1s ease" }}
+        />
+      </svg>
+      <div
+        style={{
+          position: "absolute",
+          inset: 0,
+          display: "flex",
+          flexDirection: "column",
+          alignSelf: "center",
+          alignItems: "center",
+          justifyContent: "center",
+        }}
+      >
+        <span style={{ fontSize: 14, fontWeight: 700, color: "#00e5a0", lineHeight: 1 }}>
+          {value}%
+        </span>
+        <span style={{ fontSize: 9, color: "#556677", letterSpacing: 0.5 }}>CONF</span>
+      </div>
+    </div>
+  );
+}
